# backend/src/main/java/com/aicostops/iam/api/RegisterRequest.java
package com.aicostops.iam.api;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record RegisterRequest(
        @NotBlank @Email @Size(max = 320) String email,
        @NotBlank @Size(max = 200) String displayName,
        @NotBlank @Size(min = 8, max = 200) String password) {
}