# backend/src/main/java/com/aicostops/iam/api/ResetPasswordRequest.java
package com.aicostops.iam.api;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record ResetPasswordRequest(@NotBlank String token, @NotBlank @Size(min = 8, max = 200) String newPassword) {
}