# docs/02-development/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI CostOps V1 API
  version: 1.0.0
  description: AI CostOps V1 前后端接口机器可读契约。ID 与金额均使用字符串传输。
servers:
- url: /api/v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProblemDetail:
      type: object
      required:
      - title
      - status
      - code
      - traceId
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type:
          - string
          - 'null'
        instance:
          type:
          - string
          - 'null'
        code:
          type: string
        traceId:
          type: string
        currentState:
          type:
          - string
          - 'null'
    Id:
      type: string
      pattern: ^[0-9]+$
...
    LoginRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    LoginResponse:
      type: object
      required:
      - accessToken
      - expiresIn
      - user
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
        user:
          type: object
          required:
          - id
          - displayName
          properties:
            id:
              $ref: '#/components/schemas/Id'
            displayName:
              type: string
    RegisterRequest:
      type: object
      required: [email, displayName, password]
      properties:
        email: { type: string, format: email }
        displayName: { type: string, minLength: 1, maxLength: 200 }
        password: { type: string, minLength: 8, maxLength: 200 }
    IdentityResponse:
      type: object
      required: [userId, organizationMemberId, organizationId]
      properties:
        userId: { $ref: '#/components/schemas/Id' }
        organizationMemberId: { $ref: '#/components/schemas/Id' }
        organizationId: { $ref: '#/components/schemas/Id' }
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ForgotPasswordResponse:
      type: object
      required: [accepted]
      properties:
        accepted: { type: boolean, const: true }
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token: { type: string, minLength: 1 }
        newPassword: { type: string, minLength: 8, maxLength: 200 }
    InvitationAcceptRequest:
      type: object
      required: [displayName, password]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 200 }
        password: { type: string, minLength: 8, maxLength: 200 }
    MeResponse:
      type: object
      required: [id, email, displayName, organizationId, organizationMemberId]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        email: { type: string, format: email }
        displayName: { type: string }
        organizationId: { $ref: '#/components/schemas/Id' }
        organizationMemberId: { $ref: '#/components/schemas/Id' }
...
  responses:
    Problem:
      description: 业务或请求错误
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
security:
- bearerAuth: []
paths:
  /auth/login:
    post:
      summary: 密码登录
      tags:
      - Auth
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
  /auth/register:
    post:
      summary: 公开注册（受 Feature Flag 控制）
      tags:
      - Auth
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
  /auth/refresh:
    post:
      summary: 旋转 Refresh Session
      tags:
      - Auth
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      security: []
  /auth/logout:
    post:
      summary: 注销当前 Session
      tags:
      - Auth
      responses:
        '204':
          description: Session revoked and refresh cookie cleared
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /auth/logout-all:
    post:
      summary: 注销全部 Session
      tags:
      - Auth
      responses:
        '204':
          description: Durable security version increment completed
        '401':
          $ref: '#/components/responses/Problem'
  /auth/password/forgot:
    post:
      summary: 创建密码重置 Challenge
      tags:
      - Auth
      responses:
        '202':
          description: Generic acceptance response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForgotPasswordResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
  /auth/password/reset:
    post:
      summary: 重置密码
      tags:
      - Auth
      responses:
        '204':
          description: Password changed and sessions invalidated
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
  /auth/me:
    get:
      summary: 读取当前身份与权限上下文
      tags:
      - Auth
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Problem'
  /invitations/{token}/accept:
    post:
      summary: 接受邀请
      tags:
      - Auth
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationAcceptRequest'
...
  /invitations:
    post:
      summary: 创建邀请
      tags:
      - IAM
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
      description: 权限：`USER_INVITE`。
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
        description: 调用方生成的幂等键
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...